+++ packages/odyssey-storybook/src/components/odyssey-mui/Radio/Radio.stories.tsx
@@ -18,8 +18,8 @@ import {
   FormLabel,
   Radio,
   RadioGroup,
-} from "@mui/material";
-import { visuallyHidden } from "@mui/utils";
+  visuallyHidden,
+} from "@okta/odyssey-react-mui";
 import { MuiThemeDecorator } from "../../../../.storybook/components";
 
 import RadioMdx from "./Radio.mdx";
@@ -58,7 +58,6 @@ export default {
 };
 
 const DefaultTemplate: Story = (args) => {
-  const {} = args;
   return (
     <FormControl
       component="fieldset"